ref: add tag managing mutations
